test(nimbus): move get_featmon_slugs tests into test_tasks.py
